+++ db.js
@@ -1,0 +1,144 @@
+const aws_sdk = require('aws-sdk');
+const { timeStamp } = require('console');
+
+aws_sdk.config.update({region: 'ap-south-1'});
+
+const docClient = new aws_sdk.DynamoDB.DocumentClient();
+
+function getByBookId(id) {
+    var params = {
+        TableName: "books_dev",
+        Key: {
+            book_id:parseInt(id)
+        }
+    };
+
+    return new Promise((res,rej)=>{
+        docClient.get(params, (err,response)=>{  
+            if(err){
+                rej(err);
+            }
+            if(response){
+               res(response);
+            }
+        });
+    })
+}
+
+function getByBookName(name) {
+    var params = {
+        TableName: "books_dev",
+        IndexName: "book_name-index",
+        KeyConditionExpression: "book_name = :book_name",
+    ExpressionAttributeValues: {
+        ":book_name": name
+    },
+    };
+    return new Promise((res,rej)=>{
+        docClient.query(params, (err,response)=>{  
+            if(err){
+                rej(err);
+            }
+            if(response){
+               res(response);
+            }
+        });
+    })
+}
+
+
+function getByAuthor(AuthorName) {
+    var params = {
+        TableName: "books_dev",
+        IndexName: "author-index",
+        KeyConditionExpression: "author = :author_name",
+    ExpressionAttributeValues: {
+        ":author_name": AuthorName
+    },
+    };
+    return new Promise((res,rej)=>{
+        docClient.query(params, (err,response)=>{  
+            if(err){
+                rej(err);
+            }
+            if(response){
+
+		        let arr=[];
+		        for(i=0;i<response.Items.length;i++)
+		            arr.push(response.Items[i].book_name);
+		        res((arr));
+            }
+        });
+    })
+}
+
+function buyBook(Bid) {
+    var params = {
+
+        TableName: "books_dev",
+        Key: {
+            book_id: Bid,
+
+        },
+        UpdateExpression: "set quantity = quantity - :val",
+        ConditionExpression: 'quantity > :inc',
+        ExpressionAttributeValues: {
+
+            ":val": 1,
+            ":inc": 0
+        },
+        ReturnValues: "UPDATED_NEW"
+    };
+
+    console.log("Updating the item...");
+
+    return new Promise((res, rej) => {
+
+        docClient.update(params, function (err, data) {
+            if (err) {
+                rej(err)
+            } if (data) {
+                res(data)
+            }
+        });
+    })
+
+}
+let addBook = function (valueObject) {
+    var params = {
+        TableName: "books_dev",
+        Item: {
+            book_id: parseInt(valueObject.book_id),
+            book_name: valueObject.book_name,
+            quantity: parseInt(valueObject.quantity),
+            author: valueObject.author
+
+        },
+    };
+
+
+    return new Promise((res, rej) => {
+        docClient.put(params, (err, Response) => {
+            if (err) {
+
+                rej(err)
+                //console.log("users::addBook::error - " + JSON.stringify(err, null, 2));
+            }
+            if (Response) {
+                res(Response.Item)
+                //console.log("users::addBook::success - " + JSON.stringify(Response, null, 2));
+            }
+
+        })
+    })
+
+}
+
+module.exports = {
+    getByBookId,
+	getByBookName,
+	getByAuthor,
+    buyBook,
+    docClient,
+    
+}
