+++ handler.js
@@ -1,179 +1,13 @@
 const aws_sdk = require('aws-sdk');
 const { timeStamp } = require('console');
+db = require('./db');
 
-aws_sdk.config.update({region: 'ap-south-1'});
+// aws_sdk.config.update({region: 'ap-south-1'});
 
-const docClient = new aws_sdk.DynamoDB.DocumentClient();
-
-
-function getIdByName(bookn){
-    var params = {
-        TableName : 'books_dev',
-        IndexName : 'book_name-index',
-        KeyConditionExpression : 'book_name = :book_name',
-        ExpressionAttributeValues : {
-            ':book_name' : bookn,
-        }
-    };
-
-      return new Promise((res,rej)=>{
-        docClient.query(params, (err,response)=>{  
-            if(err){
-                rej(err);
-            }
-            if(response){
-               res(response.Items[0].book_id);
-            }
-        });
-    })
-}
-
-
-
-async function getId(ctx) {
-    await getIdByName(ctx.params.bookName).then(
-        result => {
-            ctx.body=result;
-        }
-    ).catch(err=>{
-        ctx.body=err;
-    }
-     )
-}
-
-function Available(bookn){
-    var params = {
-        TableName : 'books_dev',
-        IndexName : 'book_name-index',
-        KeyConditionExpression : 'book_name = :book_name',
-        ExpressionAttributeValues : {
-            ':book_name' : bookn,
-        }
-    };
-
-      return new Promise((res,rej)=>{
-        docClient.query(params, (err,response)=>{  
-            if(err){
-                rej(err);
-            }
-            if(response){
-                if(response.Items.length!=0){
-                    res(true);
-                }
-                else
-                {
-                    res(false);
-                }
-            }
-        });
-    })
-}
-
-  function getByBookId(id) {
-    var params = {
-        TableName: "books_dev",
-        Key: {
-            book_id:parseInt(id)
-        }
-    };
-
-    return new Promise((res,rej)=>{
-        docClient.get(params, (err,response)=>{  
-            if(err){
-                rej(err);
-            }
-            if(response){
-               res(response);
-            }
-        });
-    })
-}
-
-
- function  getByBookName(name) {
-    var params = {
-        TableName: "books_dev",
-        IndexName: "book_name-index",
-        KeyConditionExpression: "book_name = :book_name",
-    ExpressionAttributeValues: {
-        ":book_name": name
-    },
-    };
-    return new Promise((res,rej)=>{
-        docClient.query(params, (err,response)=>{  
-            if(err){
-                rej(err);
-            }
-            if(response){
-               res(response);
-            }
-        });
-    })
-}
-
-  function getByAuthor(AuthorName) {
-    var params = {
-        TableName: "books_dev",
-        IndexName: "author-index",
-        KeyConditionExpression: "author = :author_name",
-    ExpressionAttributeValues: {
-        ":author_name": AuthorName
-    },
-    };
-    return new Promise((res,rej)=>{
-        docClient.query(params, (err,response)=>{  
-            if(err){
-                rej(err);
-            }
-            if(response){
-
-		let arr=[];
-		for(i=0;i<response.Items.length;i++)
-		arr.push(response.Items[i].book_name);
-		res((arr));
-              // res(response);
-            }
-        });
-    })
-}
-
-function buyBook(Bid) {
-    var params = {
-
-        TableName: "books_dev",
-        Key: {
-            book_id: Bid,
-
-        },
-        UpdateExpression: "set quantity = quantity - :val",
-        ConditionExpression: 'quantity > :inc',
-        ExpressionAttributeValues: {
-
-            ":val": 1,
-            ":inc": 0
-        },
-        ReturnValues: "UPDATED_NEW"
-    };
-
-    console.log("Updating the item...");
-
-    return new Promise((res, rej) => {
-
-        docClient.update(params, function (err, data) {
-            if (err) {
-                rej(err)
-                //console.error("Unable to update item. Error JSON:", JSON.stringify(err, null, 2));
-            } if (data) {
-                res(data)
-                //console.log("UpdateItem succeeded:", JSON.stringify(data, null, 2));
-            }
-        });
-    })
-
-}
+// const docClient = new aws_sdk.DynamoDB.DocumentClient();
 
 async function bookId(ctx) {
-    await buyBook(parseInt(ctx.params.bookId)).then(
+    await db.buyBook(parseInt(ctx.params.bookId)).then(
         result => {
             ctx.body = result;
         }
@@ -183,20 +17,10 @@ async function bookId(ctx) {
     )
 }
 
-
-
-
-
-async function isAvailable(book_name) {
-    const response = await Available(book_name);
-    return response;
-}
-
-
- async function byId(ctx) {
+async function byId(ctx) {
 
 	console.log(ctx.url);
-      await getByBookId(ctx.params.Id).then(
+      await db.getByBookId(ctx.params.Id).then(
         result => {
             ctx.body=result;
         }
@@ -205,8 +29,8 @@ async function isAvailable(book_name) {
     }
      );
 }
- async function byName(ctx) {
-      await getByBookName(ctx.params.Name).then(
+async function byName(ctx) {
+      await db.getByBookName(ctx.params.Name).then(
         result => {
             ctx.body=result;
         }
@@ -215,51 +39,23 @@ async function isAvailable(book_name) {
     }
      );
 }
- async function byAuthor(ctx) {
-      await getByAuthor(ctx.params.Author).then(
+async function byAuthor(ctx) {
+    await db.getByAuthor(ctx.params.Author).then(
         result => {
-		if(result.length==0){ ctx.body='No Data for this author';}
-	else
+		if(result.length==0){ 
+            ctx.body='No Data for this author';
+        }
+	    else
             ctx.body=result;
         }
     ).catch(err=>{
         ctx.body=err;
-    }
+        }
      );
-}
-
-let addBook = function (valueObject) {
-    var params = {
-        TableName: "books_dev",
-        Item: {
-            book_id: parseInt(valueObject.book_id),
-            book_name: valueObject.book_name,
-            quantity: parseInt(valueObject.quantity),
-            author: valueObject.author
-
-        },
-    };
-
-
-    return new Promise((res, rej) => {
-        docClient.put(params, (err, Response) => {
-            if (err) {
-
-                rej(err)
-                //console.log("users::addBook::error - " + JSON.stringify(err, null, 2));
-            }
-            if (Response) {
-                res(Response.Item)
-                //console.log("users::addBook::success - " + JSON.stringify(Response, null, 2));
-            }
-
-        })
-    })
-
 }
 
 async function callAddBook(ctx) {
-    await addBook(ctx.request.body).then(
+    await db.addBook(ctx.request.body).then(
         result => {
             ctx.body = result;
         }
@@ -270,19 +66,10 @@ async function callAddBook(ctx) {
 }
 
 module.exports = {
-    getIdByName,
-    getId,
-    isAvailable,
-    Available,
-    getByBookId,
-	getByBookName,
-	getByAuthor,
 	byId,
 	byName,
 	byAuthor,
 	callAddBook,
-	buyBook,
-	bookId,
-    
+	bookId
 }
