+++ package.json
@@ -12,6 +12,7 @@
     "aws-sdk": "^2.734.0",
     "koa": "^2.13.0",
     "koa-body": "^4.2.0",
+    "koa-logger": "^3.2.1",
     "koa-router": "^9.4.0"
   }
 }
