+++ router.js
@@ -5,7 +5,6 @@ router = new Router();
 
 const handler = require('./handler');
 
-router.get('/IdByName/:bookName', handler.getId);
 router.get('/byId/:Id',handler.byId);
 router.get('/byName/:Name',handler.byName);
 router.get('/byAuthor/:Author',handler.byAuthor);
